# docs/swagger.yaml
components:
  schemas:
...
    Item:
      properties:
        abyssJewel:
          type: boolean
        additionalProperties:
          items:
            $ref: '#/components/schemas/ItemProperty'
          type: array
          uniqueItems: false
        baseType:
          type: string
        bondedMods:
          description: PoE2 only
          items:
            type: string
          type: array
          uniqueItems: false
        colour:
          type: string
        corrupted:
          type: boolean
        cosmeticMods:
          items:
            type: string
          type: array
          uniqueItems: false
        craftedMods:
          items:
            type: string
          type: array
          uniqueItems: false
        delve:
          type: boolean
        desecrated:
          description: PoE2 only
          type: boolean
        desecratedMods:
          description: PoE2 only
          items:
            type: string
          type: array
          uniqueItems: false
        doubleCorrupted:
          description: PoE2 only
          type: boolean
        duplicated:
          type: boolean
        elder:
          type: boolean
        enchantMods:
          items:
            type: string
          type: array
          uniqueItems: false
        explicitMods:
          items:
            type: string
          type: array
          uniqueItems: false
        extended:
          $ref: '#/components/schemas/ItemExtended'
        foilVariation:
          type: integer
        foreseeing:
          type: boolean
        fractured:
          type: boolean
        fracturedMods:
          items:
            type: string
          type: array
          uniqueItems: false
        frameType:
          type: integer
        gemSockets:
          items:
            $ref: '#/components/schemas/GemSocket'
          type: array
          uniqueItems: false
        grantedSkills:
          description: PoE2 only
          items:
            $ref: '#/components/schemas/ItemProperty'
          type: array
          uniqueItems: false
        h:
          type: integer
        hybrid:
          $ref: '#/components/schemas/ItemHybrid'
        icon:
          type: string
        id:
          type: string
        identified:
          type: boolean
        ilvl:
          type: integer
        implicitMods:
          items:
            type: string
          type: array
          uniqueItems: false
        incubatedItem:
          $ref: '#/components/schemas/ItemIncubatedItem'
        influences:
          additionalProperties:
            type: boolean
          type: object
        inventoryId:
          type: string
        isRelic:
          type: boolean
        itemLevel:
          type: integer
        logbookMods:
          items:
            $ref: '#/components/schemas/ItemLogbookMod'
          type: array
          uniqueItems: false
        memoryItem:
          type: boolean
        mutated:
          type: boolean
        mutatedMods:
          items:
            type: string
          type: array
          uniqueItems: false
        name:
          type: string
        notableProperties:
          items:
            $ref: '#/components/schemas/ItemProperty'
          type: array
          uniqueItems: false
        objectiveId:
          description: filled by us and not ggg
          type: integer
        properties:
          items:
            $ref: '#/components/schemas/ItemProperty'
          type: array
          uniqueItems: false
        rarity:
          type: string
        rewards:
          items:
            $ref: '#/components/schemas/ItemReward'
          type: array
          uniqueItems: false
        runeMods:
          description: PoE2 only
          items:
            type: string
          type: array
          uniqueItems: false
        ruthless:
          type: boolean
        scourgeMods:
          items:
            type: string
          type: array
          uniqueItems: false
        searing:
          type: boolean
        shaper:
          type: boolean
        socket:
          type: integer
        socketedItems:
          items:
            $ref: '#/components/schemas/Item'
          type: array
          uniqueItems: false
        sockets:
          items:
            $ref: '#/components/schemas/ItemSocket'
          type: array
          uniqueItems: false
        split:
          type: boolean
        stackSize:
          type: integer
        support:
          type: boolean
        supportGemRequirements:
          description: PoE2 only
          items:
            $ref: '#/components/schemas/ItemProperty'
          type: array
          uniqueItems: false
        synthesised:
          type: boolean
        talismanTier:
          type: integer
        tangled:
          type: boolean
        typeLine:
          type: string
        ultimatumMods:
          items:
            $ref: '#/components/schemas/ItemUltimatumMod'
          type: array
          uniqueItems: false
        unmodifiable:
          type: boolean
        utilityMods:
          items:
            type: string
          type: array
          uniqueItems: false
        veiled:
          type: boolean
        veiledMods:
          items:
            type: string
          type: array
          uniqueItems: false
        w:
          type: integer
        weaponRequirements:
          description: PoE2 only
          items:
            $ref: '#/components/schemas/ItemProperty'
          type: array
          uniqueItems: false
        x:
          type: integer
        "y":
          type: integer
      type: object
...
    ItemProperty:
      properties:
        displayMode:
          type: integer
        icon:
          description: PoE2 only
          type: string
        name:
          type: string
        progress:
          type: number
        suffix:
          type: string
        type:
          type: integer
        values:
          items:
            items: {}
            type: array
          type: array
          uniqueItems: false
      type: object
...
    Condition:
      properties:
        field:
          $ref: '#/components/schemas/ItemField'
        operator:
          $ref: '#/components/schemas/Operator'
        value:
          type: string
      required:
      - field
      - operator
      - value
      type: object
...
    Event:
      properties:
        application_end_time:
          format: date-time
          type: string
        application_start_time:
          format: date-time
          type: string
        event_end_time:
          format: date-time
          type: string
        event_start_time:
          format: date-time
          type: string
        game_version:
          $ref: '#/components/schemas/GameVersion'
        id:
          type: integer
        is_current:
          type: boolean
        is_locked:
          type: boolean
        is_main_event:
          type: boolean
        is_public:
          type: boolean
        max_size:
          type: integer
        name:
          type: string
        patch:
          type: string
        teams:
          items:
            $ref: '#/components/schemas/Team'
          type: array
          uniqueItems: false
        uses_medals:
          type: boolean
        waitlist_size:
          type: integer
      required:
      - application_end_time
      - application_start_time
      - event_end_time
      - event_start_time
      - game_version
      - id
      - is_current
      - is_locked
      - is_main_event
      - is_public
      - max_size
      - name
      - teams
      - uses_medals
      - waitlist_size
      type: object
...
    Objective:
      properties:
        aggregation:
          $ref: '#/components/schemas/AggregationType'
        children:
          items:
            $ref: '#/components/schemas/Objective'
          type: array
          uniqueItems: false
        conditions:
          items:
            $ref: '#/components/schemas/Condition'
          type: array
          uniqueItems: false
        extra:
          type: string
        hide_progress:
          type: boolean
        id:
          type: integer
        name:
          type: string
        number_field:
          $ref: '#/components/schemas/NumberField'
        number_field_explanation:
          type: string
        objective_type:
          $ref: '#/components/schemas/ObjectiveType'
        parent_id:
          type: integer
        required_number:
          type: integer
        scoring_presets:
          items:
            $ref: '#/components/schemas/ScoringPreset'
          type: array
          uniqueItems: false
        valid_from:
          format: date-time
          type: string
        valid_to:
          format: date-time
          type: string
      required:
      - aggregation
      - children
      - conditions
      - extra
      - hide_progress
      - id
      - name
      - number_field
      - objective_type
      - parent_id
      - required_number
      - scoring_presets
      type: object
    ObjectiveCreate:
      properties:
        aggregation:
          $ref: '#/components/schemas/AggregationType'
        conditions:
          items:
            $ref: '#/components/schemas/Condition'
          type: array
          uniqueItems: false
        extra:
          type: string
        hide_progress:
          type: boolean
        id:
          type: integer
        name:
          type: string
        number_field:
          $ref: '#/components/schemas/NumberField'
        number_field_explanation:
          type: string
        objective_type:
          $ref: '#/components/schemas/ObjectiveType'
        parent_id:
          type: integer
        required_number:
          type: integer
        scoring_preset_ids:
          items:
            type: integer
          type: array
          uniqueItems: false
        valid_from:
          format: date-time
          type: string
        valid_to:
          format: date-time
          type: string
      required:
      - aggregation
      - conditions
      - name
      - number_field
      - objective_type
      - parent_id
      - required_number
      - scoring_preset_ids
      type: object
    ObjectiveValidation:
      properties:
        item:
          $ref: '#/components/schemas/Item'
        objective_id:
          type: integer
        timestamp:
          format: date-time
          type: string
      required:
      - item
      - objective_id
      - timestamp
      type: object
...
    ScoringPreset:
      properties:
        description:
          type: string
        extra:
          additionalProperties:
            type: string
          type: object
        id:
          type: integer
        name:
          type: string
        point_cap:
          type: integer
        points:
          items:
            type: number
          type: array
          uniqueItems: false
        scoring_method:
          $ref: '#/components/schemas/ScoringMethod'
      required:
      - description
      - id
      - name
      - points
      - scoring_method
      type: object
...
    Team:
      properties:
        abbreviation:
          type: string
        allowed_classes:
          items:
            type: string
          type: array
          uniqueItems: false
        color:
          type: string
        discord_role_id:
          type: string
        event_id:
          type: integer
        id:
          type: integer
        name:
          type: string
      required:
      - allowed_classes
      - event_id
      - id
      - name
      type: object
...
    ValidationRequest:
      properties:
        timeout_seconds:
          type: integer
      required:
      - timeout_seconds
      type: object
...
    AggregationType:
      enum:
      - SUM_LATEST
      - LATEST
      - EARLIEST
      - EARLIEST_FRESH_ITEM
      - MAXIMUM
      - MINIMUM
      - DIFFERENCE_BETWEEN
      - NONE
      type: string
      x-enum-varnames:
      - AggregationTypeSumLatest
      - AggregationTypeLatest
      - AggregationTypeEarliest
      - AggregationTypeEarliestFreshItem
      - AggregationTypeMaximum
      - AggregationTypeMinimum
      - AggregationTypeDifferenceBetween
      - AggregationTypeNone
...
    GameVersion:
      enum:
      - poe1
      - poe2
      type: string
      x-enum-varnames:
      - PoE1
      - PoE2
...
    NumberField:
      enum:
      - STACK_SIZE
      - PLAYER_LEVEL
      - DELVE_DEPTH
      - DELVE_DEPTH_PAST_100
      - PANTHEON
      - ASCENDANCY
      - FULLY_ASCENDED
      - BLOODLINE_ASCENDANCY_POINTS
      - BLOODLINE_ASCENDANCY
      - PLAYER_SCORE
      - HAS_RARE_ASCENDANCY_PAST_90
      - WEAPON_QUALITY
      - ARMOUR_QUALITY
      - FLASK_QUALITY
      - EVASION
      - ENERGY_SHIELD
      - ARMOUR
      - HP
      - MANA
      - FULL_DPS
      - EHP
      - INC_MOVEMENT_SPEED
      - PHYS_MAX_HIT
      - ELE_MAX_HIT
      - ATLAS_POINTS
      - INFLUENCE_EQUIPPED
      - FOULBORN_EQUIPPED
      - GEMS_EQUIPPED
      - CORRUPTED_ITEMS_EQUIPPED
      - JEWELS_WITH_IMPLICITS_EQUIPPED
      - ENCHANTED_ITEMS_EQUIPPED
      - SUBMISSION_VALUE
      - FINISHED_OBJECTIVES
      type: string
      x-enum-varnames:
      - NumberFieldStackSize
      - NumberFieldPlayerLevel
      - NumberFieldDelveDepth
      - NumberFieldDelveDepthPast100
      - NumberFieldPantheon
      - NumberFieldAscendancy
      - NumberFieldFullyAscended
      - NumberFieldBloodlineAscendancyPoints
      - NumberFieldBloodlineAscendancy
      - NumberFieldPlayerScore
      - NumberFieldHasRareAscendancyPast90
      - NumberFieldWeaponQuality
      - NumberFieldArmourQuality
      - NumberFieldFlaskQuality
      - NumberFieldEvasion
      - NumberFieldEnergyShield
      - NumberFieldArmour
      - NumberFieldHP
      - NumberFieldMana
      - NumberFieldFullDPS
      - NumberFieldEHP
      - NumberFieldIncMovementSpeed
      - NumberFieldPhysMaxHit
      - NumberFieldEleMaxHit
      - NumberFieldAtlasPoints
      - NumberFieldInfluenceEquipped
      - NumberFieldFoulbornEquipped
      - NumberFieldGemsEquipped
      - NumberFieldCorruptedItemsEquipped
      - NumberFieldJewelsWithImplicitsEquipped
      - NumberFieldEnchantedItemsEquipped
      - NumberFieldSubmissionValue
      - NumberFieldFinishedObjectives
    ObjectiveType:
      enum:
      - ITEM
      - PLAYER
      - TEAM
      - SUBMISSION
      - CATEGORY
      type: string
      x-enum-varnames:
      - ObjectiveTypeItem
      - ObjectiveTypePlayer
      - ObjectiveTypeTeam
      - ObjectiveTypeSubmission
      - ObjectiveTypeCategory
...
  securitySchemes:
    BearerAuth:
      in: header
      name: Authorization
      type: apiKey
...
info:
  contact:
    email: Liberatorist@gmail.com
    name: Liberator
  description: This is the backend API for the BPL project.
  title: BPL Backend API
  version: "2.0"
openapi: 3.1.0
paths:
...
  /events/{event_id}:
    delete:
      description: Deletes an event
      operationId: DeleteEvent
      parameters:
      - description: Event Id
        in: path
        name: event_id
        required: true
        schema:
          type: integer
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      tags:
      - event
    get:
      description: Fetches an event by id
      operationId: GetEvent
      parameters:
      - description: Event Id
        in: path
        name: event_id
        required: true
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
          description: OK
      security:
      - BearerAuth: []
      tags:
      - event
...
  /events/{event_id}/objectives:
    get:
      description: Gets all objectives for an event
      operationId: GetObjectiveTreeForEvent
      parameters:
      - description: Event Id
        in: path
        name: event_id
        required: true
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Objective'
          description: OK
      tags:
      - objective
    put:
      description: Creates a new objective
      operationId: CreateObjective
      parameters:
      - description: Event Id
        in: path
        name: event_id
        required: true
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/ObjectiveCreate'
                description: Objective to create
                summary: objectiveCreate
        description: Objective to create
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Objective'
          description: Created
      security:
      - BearerAuth: []
      tags:
      - objective
  /events/{event_id}/objectives/{id}:
    delete:
      description: Deletes an objective
      operationId: DeleteObjective
      parameters:
      - description: Event Id
        in: path
        name: event_id
        required: true
        schema:
          type: integer
      - description: Objective Id
        in: path
        name: id
        required: true
        schema:
          type: integer
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      tags:
      - objective
    get:
      description: Gets an objective by id
      operationId: GetObjective
      parameters:
      - description: Event Id
        in: path
        name: event_id
        required: true
        schema:
          type: integer
      - description: Objective Id
        in: path
        name: id
        required: true
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Objective'
          description: OK
      security:
      - BearerAuth: []
      tags:
      - objective
...
  /events/{event_id}/objectives/validations:
    get:
      description: Gets objective validations for an event
      operationId: GetObjectiveValidations
      parameters:
      - description: Event Id
        in: path
        name: event_id
        required: true
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ObjectiveValidation'
                type: array
          description: OK
      tags:
      - objective
    post:
      description: Validates item objectives for an event seeing if there are completions
        on trade
      operationId: ValidateObjectives
      parameters:
      - description: Event Id
        in: path
        name: event_id
        required: true
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/ValidationRequest'
                description: Validation request
                summary: validationRequest
        description: Validation request
        required: true
      responses:
        "204":
          description: No Content
      tags:
      - objective
...